Added xHarbour code examples
The included files are examples of my finished xHarbour code.  All code here is mine, although they rely on pre-existing libraries and modules in order to function.  The HTML_Help directory contains examples of the Jazz help environment - a locally-hosted HTML help system accessed by pressing the F1 key from any location in Jazz.  If a specific help file exists for the currently open Jazz module, the help system is designed to detect the origin location and open the related help file.  If no specific help page exists, the help pages open to the general navigation description pages.

BFVers.prg is a module whose purpose is to display the alternate versions of a base formula (a base formula is essentially the 'recipe' for a product, or 'Finished Good' in company terminology).  When the BfVers function is called, a screen displays the alternate versions (sorted by an auto-incrementing version number) in the left-hand list window, while the right-hand list window shows the ingredie